split upload controller into 2
