--- app/views/bulk_upload_logs/show.html.erb
@@ -1,1 +1,0 @@
-Find me in app/views/bulk_upload_logs/show.html.erb
