+++ app/views/bulk_upload_sales_logs/show.html.erb
@@ -1,0 +1,1 @@
+Find me in app/views/bulk_upload_sales_logs/show.html.erb
