+++ config/routes.rb
@@ -110,7 +110,7 @@ Rails.application.routes.draw do
       post "email-csv", to: "lettings_logs#email_csv"
       get "csv-confirmation", to: "lettings_logs#csv_confirmation"
 
-      resources :bulk_upload_logs, path: "bulk-upload-logs", as: "bulk_upload_lettings_logs" do
+      resources :bulk_upload_lettings_logs, path: "bulk-upload-logs" do
         collection do
           get :start
         end
@@ -135,7 +135,7 @@ Rails.application.routes.draw do
 
   resources :sales_logs, path: "/sales-logs" do
     collection do
-      resources :bulk_upload_logs, path: "bulk-upload-logs", as: "bulk_upload_sales_logs" do
+      resources :bulk_upload_sales_logs, path: "bulk-upload-logs" do
         collection do
           get :start
         end
